Move leaderboard_data into new composite_data file
